+++ agent-manager-v2/agent-manager-common/src/main/java/com/didichuxing/datachannel/agentmanager/common/bean/po/logcollecttask/CollectTaskMetricPO.java
@@ -61,7 +61,7 @@ public class CollectTaskMetricPO extends BasePO {
 
     private Integer limitRate;
 
-    private Integer controlTimeMean;
+    private Long controlTimeMean;
 
     private Integer limitTime;
 
@@ -85,7 +85,7 @@ public class CollectTaskMetricPO extends BasePO {
 
     private String logTimeStr;
 
-    private Integer controlTimeMax;
+    private Long controlTimeMax;
 
     private Integer sendCount;
 
@@ -101,7 +101,7 @@ public class CollectTaskMetricPO extends BasePO {
 
     private String collectFiles;
 
-    private Integer controlTimeMin;
+    private Long controlTimeMin;
 
     private Integer readByte;
 
@@ -335,11 +335,11 @@ public class CollectTaskMetricPO extends BasePO {
         this.limitRate = limitRate;
     }
 
-    public Integer getControlTimeMean() {
+    public Long getControlTimeMean() {
         return controlTimeMean;
     }
 
-    public void setControlTimeMean(Integer controlTimeMean) {
+    public void setControlTimeMean(Long controlTimeMean) {
         this.controlTimeMean = controlTimeMean;
     }
 
@@ -431,11 +431,11 @@ public class CollectTaskMetricPO extends BasePO {
         this.logTimeStr = logTimeStr;
     }
 
-    public Integer getControlTimeMax() {
+    public Long getControlTimeMax() {
         return controlTimeMax;
     }
 
-    public void setControlTimeMax(Integer controlTimeMax) {
+    public void setControlTimeMax(Long controlTimeMax) {
         this.controlTimeMax = controlTimeMax;
     }
 
@@ -495,11 +495,11 @@ public class CollectTaskMetricPO extends BasePO {
         this.collectFiles = collectFiles;
     }
 
-    public Integer getControlTimeMin() {
+    public Long getControlTimeMin() {
         return controlTimeMin;
     }
 
-    public void setControlTimeMin(Integer controlTimeMin) {
+    public void setControlTimeMin(Long controlTimeMin) {
         this.controlTimeMin = controlTimeMin;
     }
